Add support for Flymaster NAV
This patch adds support for the Flymaster NAV (newer generation of
Flymaster B1NAV) to flightrecorder.

B1NAV and NAV mostly use the same serial protocol, but there are
a few quirks:

- The NAV indexes tracks from 1 to count in the PFMDNL,LST-response,
while the B1NAV indexes from 0 to count-1. This patch adjusts the
track index by one if the NAV is detected.

- The NAV sends a complete 4-byte response to signal end of binary
mode (0xA3A3), unlike the B1NAV which only contains the 2-byte ID.
The length field is always set to 0 for this packet.